requirements.txt
aioboto3==13.2.0
aiobotocore==2.15.2
aiofiles==24.1.0
aiohttp==3.9.5
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
azure-ai-inference==1.0.0b9
azure-core==1.30.1
azure-core-tracing-opentelemetry==1.0.0b11
azure-identity==1.16.0
azure-monitor-opentelemetry==1.6.0
azure-monitor-opentelemetry-exporter==1.0.0b27
azure-storage-blob==12.24.0
beautifulsoup4==4.12.3
boto3==1.35.36
botocore==1.35.36
build==1.2.1
cachetools==5.5.2
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.11.2
cryptography==42.0.7
dataclasses-json==0.6.7
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
environs==11.0.0
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
-e git+https://github.com/jawad280/rag-google-hd.git@69c8a2510d68184c75c543d7c2e94c18669548e9#egg=fastapi_app&subdirectory=src
fastavro==1.9.7
filelock==3.14.0
fixedint==0.1.6
frozenlist==1.4.1
fsspec==2024.10.0
geographiclib==2.0
geopy==2.4.1
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.39.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.26.1
identify==2.5.36
idna==3.7
importlib_metadata==7.1.0
isodate==0.6.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-aws==0.2.6
langchain-community==0.3.5
langchain-core==0.3.15
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.139
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.93.0
openai-messages-token-helper==0.1.4
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-dbapi==0.46b0
opentelemetry-instrumentation-django==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-flask==0.46b0
opentelemetry-instrumentation-logging==0.46b0
opentelemetry-instrumentation-psycopg2==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-sqlalchemy==0.46b0
opentelemetry-instrumentation-urllib==0.46b0
opentelemetry-instrumentation-urllib3==0.46b0
opentelemetry-instrumentation-wsgi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.3
packaging==24.0
pandas==2.2.2
parameterized==0.9.0
pgvector==0.2.5
pillow==10.3.0
pip-tools==7.4.1
platformdirs==4.2.2
portalocker==2.8.2
pre-commit==3.7.1
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
pyproject_hooks==1.1.0
pythainlp==5.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rank-bm25==0.2.2
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9.1
ruff==0.4.8
s3transfer==0.10.3
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
slack==0.0.2
slack-sdk==3.23.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.30
starlette==0.37.2
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.20.1
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.22.0
websockets==12.0
wheel==0.44.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
